Extra fix for #33 and https://github.com/zyedidia/micro/issues/2349
Hopefully, this will close the case shut once and for all.

Yes, I recognize that these notice messages are excessive and I know you are tempted to get rid of this feature, but we can't let the small portion of Linux users who can't read a basic straightforward error message get in the way of helping so many people who love this feature.